+++ src/compas_cra/equilibrium/__init__.py
@@ -1,9 +1,12 @@
 from .cra_pyomo import cra_solve
 from .cra_penalty_pyomo import cra_penalty_solve
 from .rbe_pyomo import rbe_solve
+from .rbe_robust import plot_rbe_robust_results
 from .rbe_robust import RobustForceResult
 from .rbe_robust import rbe_robust_sample
 from .rbe_robust import rbe_robust_support
+from .rbe_robust import rbe_robust_support_dual
+from .rbe_robust import rbe_robust_support_primal
 from .cra_helper import (
     equilibrium_setup,
     friction_setup,
@@ -32,6 +35,9 @@ __all__ = [
     "rbe_solve",
     "rbe_robust_sample",
     "rbe_robust_support",
+    "rbe_robust_support_primal",
+    "rbe_robust_support_dual",
+    "plot_rbe_robust_results",
     "RobustForceResult",
     "equilibrium_setup",
     "friction_setup",